requirements.txt
discord.py
paramiko